Address quantization bug when writing images
I assumed that the principal components produced by RandomizedPCA had values in
(0..1).  This assumption was incorrect and led to gray colors in the eigenfaces
being swamped out to full-black or full-white when written to disk using
skimage.io.imsave. Switching to matplotlib.image.imsave fixes this problem.

Reported by @PetterS:
- https://github.com/c-w/github-data-challenge-2014/commit/049487e#commitcomment-8585362
- https://github.com/c-w/github-data-challenge-2014/commit/049487e#commitcomment-8605289